src/web/requirements.txt
appdirs==1.4.4
apturl==0.5.2
bar==0.2.1
bcrypt==3.1.7
blinker==1.4
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
DateTime==4.3
dbus-python==1.2.16
defer==1.0.6
dill==0.3.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==0.4.15
httplib2==0.14.0
idna==2.8
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
MarkupSafe==1.1.1
matplotlib==3.4.1
menulibre==2.2.1
monotonic==1.5
mugshot==0.4.2
multiprocess==0.70.11.1
netifaces==0.10.4
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
pandas==1.2.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyserial==3.5
python-3parclient==4.2.11
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.11
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
utc==0.0.3
virtualenv==20.4.3
wadllib==1.3.3
Werkzeug==1.0.1
xcffib==0.8.1
xkit==0.0.0
zope.interface==5.3.0
